+++ documents/docs/zh-CN/changelog.md
@@ -33,6 +33,8 @@
 
 [新增] excel导出 支持导出当前页数据，配置 `excel.curPage = true` [四、分页情况下只导出当前页数据](#/zh-CN/component/excel/basic)
 
+[修复] excel导出 修复右对其失效问题
+
 ### **1.4.4** <small>`2019-11-12`</small>
 
 [修复] excel导出 支持 `method`、`contentType` 参数