+++ documents/docs/zh-CN/changelog.md
@@ -31,9 +31,11 @@
 
 [修复] excel导出 修复数据 0 导出为空的问题
 
-[新增] excel导出 支持导出当前页数据，配置 `excel.curPage = true` [四、分页情况下只导出当前页数据](#/zh-CN/component/excel/basic)
+[增强] excel导出 支持导出当前页数据，配置 `excel.curPage = true` [四、分页情况下只导出当前页数据](#/zh-CN/component/excel/basic)
 
 [修复] excel导出 修复右对其失效问题
+
+[增强] 子表 子表宽度支持铺满父表 [7. 子表宽度铺满父表](#/zh-CN/component/child/show)
 
 ### **1.4.4** <small>`2019-11-12`</small>