+++ documents/docs/zh-CN/changelog.md
@@ -37,6 +37,8 @@
 
 [增强] 子表 子表宽度支持铺满父表 [7. 子表宽度铺满父表](#/zh-CN/component/child/show)
 
+[增强] 子表 弹窗方式支持通过 `layerOption` 自定义弹窗的配置 [3.弹窗方式](#/zh-CN/component/child/show)
+
 ### **1.4.4** <small>`2019-11-12`</small>
 
 [修复] excel导出 支持 `method`、`contentType` 参数